requirements.txt
argon2-cffi==21.1.0
asgiref==3.4.1
attrs==21.2.0
autopep8==1.5.7
backcall==0.2.0
backports.entry-points-selectable==1.1.0
bleach==4.1.0
certifi==2021.10.8
cffi==1.14.6
charset-normalizer==2.1.0
colorama==0.4.4
cycler==0.10.0
debugpy==1.4.3
decorator==5.1.0
defusedxml==0.7.1
demjson==2.2.4
distlib==0.3.3
Django==3.2.7
django-taggit==1.2.0
djangorestframework==3.12.4
entrypoints==0.3
filelock==3.3.0
idna==3.3
ipykernel==6.4.1
ipython==7.28.0
ipython-genutils==0.2.0
jedi==0.18.0
Jinja2==3.0.1
jmespath==1.0.1
joblib==1.1.0
jsonschema==3.2.0
jupyter-client==7.0.3
jupyter-core==4.8.1
jupyterlab-pygments==0.1.2
kiwisolver==1.3.2
MarkupSafe==2.0.1
matplotlib==3.4.3
matplotlib-inline==0.1.3
mistune==0.8.4
nbclient==0.5.4
nbconvert==6.2.0
nbformat==5.1.3
nest-asyncio==1.5.1
nester==0.0.1
notebook==6.4.4
numpy==1.21.2
packaging==21.0
pandas==1.3.3
pandocfilters==1.5.0
parso==0.8.2
pickleshare==0.7.5
Pillow==8.3.2
pipenv==2021.5.29
platformdirs==2.4.0
prometheus-client==0.11.0
prompt-toolkit==3.0.20
pycodestyle==2.7.0
pycparser==2.20
pygame==2.0.3
Pygments==2.10.0
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil==2.8.2
pytz==2021.1
pywin32==301
pywinpty==1.1.4
pyzmq==22.3.0
requests==2.28.1
router==0.1
routers==0.10.1
scikit-learn==1.0
scipy==1.7.1
Send2Trash==1.8.0
six==1.16.0
sklearn==0.0
sqlparse==0.4.2
terminado==0.12.1
testpath==0.5.0
threadpoolctl==3.0.0
toml==0.10.2
tornado==6.1
traitlets==5.1.0
urllib3==1.26.10
virtualenv==20.8.1
virtualenv-clone==0.5.7
wcwidth==0.2.5
webencodings==0.5.1
xlrd==2.0.1
XlsxWriter==3.0.3
xlwt==1.3.0
